refactor : CommentStore -> CommentsList 및 첫글자 소문자로 변경
