+++ src/app/_store/commentStore.ts
@@ -1,13 +1,13 @@
 import { CommentListType } from '@/type';
 import { create } from 'zustand';
 
-interface CommentState {
+interface CommentListState {
   comments: CommentListType[];
   setComments: (comments: CommentListType[]) => void;
   reset: () => void;
 }
 
-export const useCommentStore = create<CommentState>((set) => ({
+export const useCommentListStore = create<CommentListState>((set) => ({
   comments: [],
   setComments: (comments: CommentListType[]) => set({ comments }),
   reset: () => set({ comments: [] }),
